+++ lib/unisec.rb
@@ -2,6 +2,7 @@
 
 require 'unisec/version'
 
+require 'unisec/bidi'
 require 'unisec/confusables'
 require 'unisec/hexdump'
 require 'unisec/properties'